requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pyyaml
pytest
rich
streamlit
modal
